+++ dwtools/Sounds_to_DTW.cpp
@@ -1,6 +1,6 @@
 /* Sounds_to_DTW.cpp
  *
- * Copyright (C) 2012, 2015 David Weenink
+ * Copyright (C) 2012-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -26,7 +26,7 @@
 
 autoDTW Sounds_to_DTW (Sound me, Sound thee, double analysisWidth, double dt, double band, int slope) {
 	try {
-        long numberOfCoefficients = 12;
+        integer numberOfCoefficients = 12;
         double fmin_mel = 100, df_mel = 100, fmax_mel = 0.0;
 		autoMFCC mfcc_me = Sound_to_MFCC (me, numberOfCoefficients, analysisWidth, dt, fmin_mel, fmax_mel, df_mel);
 		autoMFCC mfcc_thee = Sound_to_MFCC (thee, numberOfCoefficients, analysisWidth, dt, fmin_mel, fmax_mel, df_mel);
